Path command precision error is fixed by switching to float from BigDecimal

+++ shapes/src/main/java/eo/view/batterymeter/shapes/PathCommand.kt
@@ -1,8 +1,6 @@
 package eo.view.batterymeter.shapes
 
-import java.math.BigDecimal
-
-class PathCommand(val command: Char, vararg val parameters: BigDecimal) {
+class PathCommand(val command: Char, vararg val parameters: Float) {
     override fun toString(): String {
         return "$command ${parameters.joinToString(" ")}"
     }
